Refactor project skeleton generation into separate deterministic module
